fix missing-seed-file bug, explanatory comments

+++ run/run-lf-model.bash
@@ -21,11 +21,17 @@ function run_ID () {
 	STARTING_YEAR="${STARTING_YEAR:=2020}"
 
 	if [[ "${USE_SEED_FILE}" = "true" ]] && [[ -n $( realpath "${SEED}" 2>/dev/null ) ]] ; then
+		echo "--> NOTE using seed file ${SEED}"
 		SEED_ARG="${SEED}"
-		echo "--> WARNING not using seed file ${SEED} as it doesn't exist"
 	else
-		SEED_ARG=1
+		echo "--> WARNING not using seed file ${SEED} as it doesn't exist"
+		# NB this is an empty string because transfil_N seems to fail on '' or "" in the -g arg
+		SEED_ARG=""
 	fi
+
+	# transfil_N wants 0 or 1 arguments for true/false
+	[[ "${OUTPUT_ENDGAME}" = "true" ]] && OUTPUT_ENDGAME_ARG=1 || OUTPUT_ENDGAME_ARG=0
+	[[ "${REDUCE_IMP_VIA_XML}" = "true" ]] && REDUCE_IMP_VIA_XML_ARG=1 || REDUCE_IMP_VIA_XML_ARG=0
 
 	echo "== making result directory ${RESULTS} for ID ${id}"
 
@@ -37,8 +43,8 @@ function run_ID () {
 		-s "${SCENARIO}" \
 		-p "${PARAMS}" \
 		-t "${TIMESTEP}" \
-		-e "${OUTPUT_ENDGAME}" \
-		-x "${REDUCE_IMP_VIA_XML}" \
+		-e "${OUTPUT_ENDGAME_ARG}" \
+		-x "${REDUCE_IMP_VIA_XML_ARG}" \
 		-g "${SEED_ARG}" \
 		-o "${RESULTS}" \
 		-n "${POP_DISTRIBUTION_FILE}" \
@@ -48,8 +54,12 @@ function run_ID () {
 	echo "== combining output files for IHME & NTDMC using output folder ${output_folder_name}"
 	( time do_file_combinations "${id}" "${output_folder_name}" "${RESULTS}" ) 2>&1
 
-	echo "== clearing out model 'result' files"
-	rm -rf "${RESULTS}"
+	if [[ "${KEEP_INTERMEDIATE_RESULTS}" = "false" ]] ; then
+		echo "== clearing out model 'result' files"
+		rm -rf "${RESULTS}"
+	else
+		echo "== leaving model 'result' files in place in ${RESULTS}"
+	fi
 }
 
 function do_file_combinations () {
@@ -58,7 +68,7 @@ function do_file_combinations () {
 	output_folder_name=${2}
     result_folder_path=${3}
 
-	for scen_iu in $( xmllint --xpath "/Model/ScenarioList/scenario/@name" <( tail -n +2 ${SCENARIO_ROOT}/scenariosNoImp${id}.xml ) | sed 's/name="\([^"]*\)"/\1/g' ) ; do
+	for scen_iu in $( xmllint --xpath "/Model/ScenarioList/scenario/@name" <( tail -n +2 ${SCENARIO_ROOT}/${SCENARIO_FILE_STEM}${id}.xml ) | sed 's/name="\([^"]*\)"/\1/g' ) ; do
 
 		scen=$( echo $scen_iu | cut -f 1 -d _ )
 		iu=$( echo $scen_iu | cut -f 2 -d _ )
